\name{RegOfPkg-package}
\alias{RegOfPkg-package}
\alias{RegOfPkg}
\docType{package}
\title{
Install Fund Approved R Packages
}
\description{
Install Fund Approved R Packages
}
\details{
\tabular{ll}{
Package: \tab RegOfPkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-26\cr
License: \tab What license is it under?\cr
}
\code{InstallApprovedPackages(pkgs)}
}
\author{
}
\references{
}
\examples{
InstallApprovedPackages(c("abc","zoo"))
}
